+++ packages/aurelia-chart/src/attributes/chart-attribute.ts
@@ -1,53 +1,58 @@
-import { inject, customAttribute, bindable, bindingMode } from 'aurelia-framework';
-import { ModelObserver } from '../observers/model-observer';
+import { BindingMode, bindable, computed, customAttribute, resolve, watch } from 'aurelia';
 import { Chart, ChartOptions, ChartData, ChartConfiguration, ChartType } from 'chart.js';
 
 @customAttribute('chart')
-@inject(Element, ModelObserver)
 export class ChartAttribute {
-  constructor(private element: HTMLCanvasElement, private modelObserver: ModelObserver) { }
-
+  private element = resolve(Element) as HTMLCanvasElement;
   activeChart?: Chart;
-  private chartData: ChartConfiguration;
+  private chartData?: ChartConfiguration;
+  private refreshTimeout?: ReturnType<typeof setTimeout>;
 
   @bindable
-  type: ChartType;
+  type: ChartType = 'line';
   typeChanged() {
+    if (!this.chartData) {
+      return;
+    }
     this.chartData.type = this.type;
     if (this.isObserving) {
       this.refreshChart();
-      this.modelObserver.unsubscribe();
-      this.subscribeToChanges();
     }
   }
 
   @bindable
-  data: ChartData;
+  data: ChartData = { datasets: [] };
   dataChanged() {
-    this.chartData.data = this.data;
-    if (this.isObserving) {
-      this.refreshChart();
-      this.modelObserver.unsubscribe();
-      this.subscribeToChanges();
+    if (!this.chartData) {
+      return;
     }
+    this.chartData.data = this.data;
   }
 
   @bindable
-  shouldUpdate: boolean | string;
+  shouldUpdate: boolean | string = false;
 
   private get isObserving() {
     return this.shouldUpdate === true || this.shouldUpdate === 'true';
   }
 
+  @computed({ deps: ['data'], deep: true })
+  get observedData(): ChartData | undefined {
+    return this.data == null ? undefined : { ...this.data };
+  }
+
+  @watch('observedData')
+  protected observedDataChanged() {
+    if (this.isObserving) {
+      this.scheduleRefresh();
+    }
+  }
+
   @bindable
   throttle?: number;
 
-  @bindable({ defaultBindingMode: bindingMode.twoWay })
+  @bindable({ mode: BindingMode.twoWay })
   nativeOptions: ChartOptions = {};
-
-  bind() {
-    // prevent initial changed handlers call
-  }
 
   attached() {
     this.chartData = {
@@ -59,15 +64,12 @@ export class ChartAttribute {
     this.activeChart = new Chart(this.element, this.chartData);
     this.nativeOptions = this.activeChart.options;
     this.refreshChart();
-
-    if (this.isObserving) {
-      this.subscribeToChanges();
-    }
   }
 
-  detached() {
-    if (this.isObserving) {
-      this.modelObserver.unsubscribe();
+  detaching() {
+    if (this.refreshTimeout !== undefined) {
+      clearTimeout(this.refreshTimeout);
+      this.refreshTimeout = undefined;
     }
 
     this.activeChart?.destroy();
@@ -81,8 +83,20 @@ export class ChartAttribute {
     }
   };
 
-  subscribeToChanges() {
-    this.modelObserver.throttle = this.throttle ?? 100;
-    this.modelObserver.observe(this.data, this.refreshChart);
+  private scheduleRefresh() {
+    const throttle = this.throttle ?? 100;
+    if (throttle <= 0) {
+      this.refreshChart();
+      return;
+    }
+
+    if (this.refreshTimeout === undefined) {
+      this.refreshTimeout = setTimeout(() => {
+        this.refreshTimeout = undefined;
+        if (this.isObserving) {
+          this.refreshChart();
+        }
+      }, throttle);
+    }
   }
 }